pulled identical device provider code into a shared base class
